pyproject.toml
[project]
name = "disparaai"
version = "0.1.0"
description = "WhatsApp Bulk Messaging System with AI-powered copywriting using AGNO framework"
authors = [
    { name = "Victor", email = "victor@disparaai.com" }
]
requires-python = ">=3.11"
dependencies = [
    # Core AGNO Framework
    "agno>=1.7.6",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    # Database & Storage
    "psycopg[binary]>=3.1.8",
    "sqlalchemy>=2.0.0",
    # WhatsApp & API Integration
    "httpx>=0.24.0",
    # File Processing
    "polars>=0.20.0",
    "python-dotenv>=1.0.0",
    "phonenumbers>=8.13.0",
    # AI & Content Generation
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    # Utilities
    "pydantic>=2.0.0",
    "pillow>=11.3.0",
    "requests>=2.32.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"
